Move common internal classes, functions, constants to _common_util module
- Rename index_constants.py to _common_util.py
- Move internal constants to _common_util.py
- Move common internal functions to _common_util.py
- Move common internal classes (Code) to _common_util.py
- Update modules throughout library to import from _common_util
- Update tests to import from _common_util when appropriate
