+++ src/cloudant/design_document.py
@@ -19,8 +19,7 @@ from ._2to3 import iteritems_
 from .document import Document
 from .view import View, QueryIndexView
 from .error import CloudantArgumentError, CloudantException
-
-QUERY_LANGUAGE = 'query'
+from ._common_util import QUERY_LANGUAGE
 
 class DesignDocument(Document):
     """
@@ -66,10 +65,8 @@ class DesignDocument(Document):
         (JSON query index) using this method is made.
 
         :param str view_name: Name used to identify the View.
-        :param str map_func: Javascript map function.  Can also be a
-            :class:`~cloudant.view.Code` object.
+        :param str map_func: Javascript map function.
         :param str reduce_func: Optional Javascript reduce function.
-            Can also be a :class:`~cloudant.view.Code` object.
         """
         if self.get_view(view_name) is not None:
             msg = "View {0} already exists in this design doc".format(view_name)
@@ -93,10 +90,8 @@ class DesignDocument(Document):
         QueryIndexView (JSON query index) using this method is made.
 
         :param str view_name: Name used to identify the View.
-        :param str map_func: Javascript map function.  Can also be a
-            :class:`~cloudant.view.Code` object.
+        :param str map_func: Javascript map function.
         :param str reduce_func: Optional Javascript reduce function.
-            Can also be a :class:`~cloudant.view.Code` object.
         """
         view = self.get_view(view_name)
         if view is None:
