+++ src/cloudant/document.py
@@ -23,7 +23,6 @@ from requests.exceptions import HTTPError
 from ._2to3 import url_quote, url_quote_plus
 from .error import CloudantException
 
-
 class Document(dict):
     """
     Encapsulates a JSON document.  A Document object is instantiated with a
