+++ src/cloudant/index.py
@@ -20,10 +20,10 @@ import posixpath
 import json
 
 from ._2to3 import STRTYPE, iteritems_
-from .index_constants import JSON_INDEX_TYPE
-from .index_constants import TEXT_INDEX_TYPE
-from .index_constants import SPECIAL_INDEX_TYPE
-from .index_constants import TEXT_INDEX_ARGS
+from ._common_util import JSON_INDEX_TYPE
+from ._common_util import TEXT_INDEX_TYPE
+from ._common_util import SPECIAL_INDEX_TYPE
+from ._common_util import TEXT_INDEX_ARGS
 from .error import CloudantArgumentError, CloudantException
 
 class Index(object):
