--- src/cloudant/index_constants.py
@@ -1,23 +1,0 @@
-#!/usr/bin/env python
-# Copyright (c) 2015 IBM. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Module containing all index related constants
-"""
-
-JSON_INDEX_TYPE = 'json'
-TEXT_INDEX_TYPE = 'text'
-SPECIAL_INDEX_TYPE = 'special'
-
-TEXT_INDEX_ARGS = {'fields': list, 'default_field': dict, 'selector': dict}
